+++ src/main/kotlin/dev/d1s/linda/exception/badRequest/impl/utmParameter/IllegalUtmParametersException.kt
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-package dev.d1s.linda.exception.alreadyExists.impl
+package dev.d1s.linda.exception.badRequest.impl.utmParameter
 
-import dev.d1s.advice.domain.ErrorResponseData
-import dev.d1s.advice.exception.HttpStatusException
-import dev.d1s.linda.constant.error.ERROR_UTM_PARAMETER_ALREADY_EXISTS
-import org.springframework.http.HttpStatus
+import dev.d1s.linda.constant.error.ILLEGAL_UTM_PARAMETERS_ERROR
+import dev.d1s.linda.entity.utmParameter.UtmParameter
+import dev.d1s.linda.exception.badRequest.BadRequestException
 
-object UtmParameterAlreadyExistsException : HttpStatusException(
-    ErrorResponseData(HttpStatus.UNPROCESSABLE_ENTITY, ERROR_UTM_PARAMETER_ALREADY_EXISTS)
+class IllegalUtmParametersException(
+    utmParameters: Set<UtmParameter>
+) : BadRequestException(
+    ILLEGAL_UTM_PARAMETERS_ERROR.format(utmParameters)
 )
